+++ src/utils/Watcher.js
@@ -8,7 +8,7 @@ import gitignore from 'parse-gitignore'
 // Internal Dependencies
 import { CONFIG_FILE_NAME, DEV_FOLDER_NAME } from '../config/CLI.js'
 import { SETUP_FOLDER_NAME } from '../config/Components.js'
-import { JSON_EXTENSION, LIQUID_EXTENSION, SCRIPT_EXTENSIONS, STYLE_EXTENSIONS } from './ComponentFilesUtils.js'
+import { JSON_EXTENSION, LIQUID_EXTENSION, SCRIPT_EXTENSIONS, STYLE_EXTENSIONS } from './componentFilesFactory.js'
 import { debug, error, logWatcherAction } from './LoggerUtils.js'
 
 export const ChangeType = {
